chore: update deploy config
