+++ README.md
@@ -16,3 +16,6 @@
   - TypeORM 설정, 마이그레이션 추가
   - Seeder 추가
   - JWT 가드 구현
+- **2025-08-16**
+  - Cool SMS API 통합 완료
+  - 전화번호 인증 SMS 발송 기능 구현
